pyproject.toml
[build-system]
requires = ['setuptools>=40.8.0']
build-backend = 'setuptools.build_meta'

[project]
name = "django-scidatacontainer"
version = "0.2.1"
authors = [
           {name="Sven Kleinert", email="sven.kleinert@phoenixd.uni-hannover.de"}
           ]
description = "A Django webserver to handle SciDataConatiner."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.1",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "Django >= 4.1.3",
    "django-filter >= 22.1",
    "django-guardian >= 2.4.0",
    "djangorestframework >= 3.14.0",
    "django-rest-knox >= 4.2.0",
    "h5py >= 3.9.0",
    "packaging >= 23.0",
    "python-magic >= 0.4.27",
    "iso8601 >= 2.0.0",
    "SciDataContainer >= 1.0.0",
]

[projects.urls]
Homepage = "https://github.com/SciDataContainer/django-scidatacontainer"
Documentation = "https://django-scidatacontainer.readthedocs.io/en/latest"
"Bug Tracker" = "https://github.com/SciDataContainer/django-scidatacontainer/issues"
